+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,6 +6,12 @@
   * Updated to support a peer dependency of react-intl v2-v5
   * Updated to use `injectIntl` instead of `FormattedMessage` for react-intl v2-v5 compatibility
 
+* Added
+  * Added `initialTimeZone` prop which takes a valid timezone string. This enables date-time-picker set to a particular timezone. The default timezone is set to the local/browser timezone.
+
+* Fixed
+  * Fixed flaky test failure due to loading unintended test example.
+
 ## 4.53.0 - (November 24, 2020)
 
 * Changed
@@ -41,8 +47,6 @@
 * Changed
   * `onClickOutside` prop has been deprecated, will resolve to `onRequestClose`
   * Updated wdio snapshots due to changes in date-picker and time-input
-
-* Changed
   * Update `timeVariant` prop description to include `es` in locales supporting only 24 hour format.
 
 ## 4.46.0 - (September 10, 2020)