+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 5.28.0 - (November 10, 2020)
 
 * Changed